Fix operational defects: DB schema, type mismatch, total_years bug
- RaceSpecificAccuracyRepository: fix type mismatch Long→Integer to
  match entity (prevented Spring Boot startup in strict mode)
- race_specific_analyzer: fix total_years hardcoded 0 → pass actual
  years arg to save_prediction_pattern(); add years= param to fn
- schema.sql: remove duplicate top5hit columns (kept top5_hit which
  is the correct JPA snake_case mapping); add UNIQUE constraints on
  race_specific_result(race_name, horse_name) and
  race_specific_prediction(race_name) to prevent duplicate records
- deploy/migrate_v2.sql: migration script for existing databases

+++ src/main/java/com/faceprediction/repository/RaceSpecificAccuracyRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.faceprediction.entity.RaceSpecificAccuracy;
 
-public interface RaceSpecificAccuracyRepository extends JpaRepository<RaceSpecificAccuracy, Long> {
+public interface RaceSpecificAccuracyRepository extends JpaRepository<RaceSpecificAccuracy, Integer> {
 
     /** 1着的中率を出すための集計：[totalRaces, winHits] */
     @Query("SELECT COUNT(DISTINCT a.raceName), " +
